# openapi.yaml
...
openapi: 3.0.3
info:
  title: EyeX API
  version: 1.2.0
  description: >-
    API de temas de pantalla para daltonismo y baja vision. Devuelve paletas completas
    listas para aplicar a una interfaz e incluye simulacion de vision de color con severidad
    continua mediante el modelo Machado, Oliveira & Fernandes (2009). Los endpoints v1 y
    sus payloads principales mantienen paridad entre Go, PHP, TypeScript/Node y Java. El
    servidor Go oficial agrega mensajes de error segun Accept-Language, X-API-Key opcional,
    rate limiting, HTTPS obligatorio en produccion, timeout y metricas Prometheus.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Entorno local
paths:
  /api/v1/theme/types:
    get:
      summary: Lista los tipos de tema disponibles
      operationId: listThemeTypes
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ApiKey'
      responses:
        '200':
          description: Tipos disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypesResponse'
              example:
                types: [normal, protanopia, deuteranopia, tritanopia, achromatopsia, low_vision]
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
  /api/v1/theme/{type}:
    get:
      summary: Obtiene una paleta completa de tema
      operationId: getTheme
      parameters:
        - name: type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ThemeType'
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [mild, moderate, severe]
          description: Intensidad de la adaptacion. Si se omite, el contrato base conserva su comportamiento original.
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [dark, light]
          description: Variante oscura o clara.
        - name: high_contrast
          in: query
          required: false
          schema:
            type: boolean
          description: Refuerza el contraste manteniendo el tipo seleccionado.
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/ApiKey'
      responses:
        '200':
          description: Tema calculado
          headers:
            ETag:
              description: Identificador del cuerpo para revalidacion HTTP.
              schema: { type: string }
            Cache-Control:
              description: Politica de cache del tema.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        '304':
          description: El tema no cambio respecto del ETag enviado.
        '400': { $ref: '#/components/responses/BadRequest' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
  /api/v1/theme/custom:
    post:
      summary: Adapta una paleta de marca
      operationId: createCustomTheme
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomThemeRequest'
      responses:
        '200':
          description: Paleta de marca adaptada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
  /api/v1/test/suggest:
    post:
      summary: Sugiere de forma orientativa un tipo de tema
      description: El resultado es una guia y no constituye un diagnostico medico.
      operationId: suggestThemeType
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickTestRequest'
      responses:
        '200':
          description: Sugerencia orientativa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuickTestResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
  /api/v1/simulate:
    post:
      summary: Simula la percepcion de un color para una deficiencia de vision cromatica
      description: >-
        Aplica las matrices de Machado, Oliveira & Fernandes (2009) en RGB lineal. La severidad
        es continua entre 0 y 1; se interpolan linealmente las matrices tabuladas cada 0.1.
      operationId: simulateColor
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateRequest'
            example:
              hex: '#FF0000'
              type: protanopia
              severity: 0.65
      responses:
        '200':
          description: Color simulado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateResponse'
              example:
                original: '#FF0000'
                simulated: '#A05A00'
                type: protanopia
                severity: 0.65
                model: machado-2009
        '400': { $ref: '#/components/responses/BadRequest' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
  /api/v1/simulate/batch:
    post:
      summary: Simula varios colores con el mismo tipo y severidad
      operationId: simulateColorBatch
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ApiKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateBatchRequest'
            example:
              colors: ['#FF0000', '#00FF00', '#0000FF']
              type: deuteranopia
              severity: 0.5
      responses:
        '200':
          description: Colores simulados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateBatchResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '426': { $ref: '#/components/responses/HttpsRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/Timeout' }
...
components:
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: Idioma de los mensajes de error. Se admiten es y en; espanol es el valor predeterminado.
      schema:
        type: string
        example: es
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag recibido en una consulta anterior del mismo tema.
      schema: { type: string }
    ApiKey:
      name: X-API-Key
      in: header
      required: false
      description: Clave opcional configurada por el operador para omitir el limite publico de solicitudes.
      schema: { type: string }
  schemas:
    ThemeType:
      type: string
      enum: [normal, protanopia, deuteranopia, tritanopia, achromatopsia, low_vision]
    HexColor:
      type: string
      pattern: '^#[0-9A-Fa-f]{6}$'
      example: '#1E1E1E'
    Palette:
      type: object
      additionalProperties: false
      required: [background, surface, text, primary, secondary, error, success]
      properties:
        background: { $ref: '#/components/schemas/HexColor' }
        surface: { $ref: '#/components/schemas/HexColor' }
        text: { $ref: '#/components/schemas/HexColor' }
        primary: { $ref: '#/components/schemas/HexColor' }
        secondary: { $ref: '#/components/schemas/HexColor' }
        error: { $ref: '#/components/schemas/HexColor' }
        success: { $ref: '#/components/schemas/HexColor' }
    ThemeResponse:
      type: object
      additionalProperties: false
      required: [type, palette, contrast_ok]
      properties:
        type: { $ref: '#/components/schemas/ThemeType' }
        palette: { $ref: '#/components/schemas/Palette' }
        contrast_ok:
          type: boolean
          description: true cuando text alcanza al menos 4.5:1 contra background y surface.
    TypesResponse:
      type: object
      additionalProperties: false
      required: [types]
      properties:
        types:
          type: array
          items: { $ref: '#/components/schemas/ThemeType' }
    CustomThemeRequest:
      type: object
      additionalProperties: false
      required: [type, palette]
      properties:
        type: { $ref: '#/components/schemas/ThemeType' }
        severity:
          type: string
          enum: [mild, moderate, severe]
        mode:
          type: string
          enum: [dark, light]
        high_contrast:
          type: boolean
          default: false
        palette: { $ref: '#/components/schemas/Palette' }
    QuickTestAnswers:
      type: object
      additionalProperties: false
      required: [reds_look_darker, green_brown_confusion, blue_yellow_confusion, colors_look_gray]
      properties:
        reds_look_darker: { type: boolean }
        green_brown_confusion: { type: boolean }
        blue_yellow_confusion: { type: boolean }
        colors_look_gray: { type: boolean }
    QuickTestRequest:
      type: object
      additionalProperties: false
      required: [answers]
      properties:
        answers: { $ref: '#/components/schemas/QuickTestAnswers' }
    QuickTestResponse:
      type: object
      additionalProperties: false
      required: [suggested_type, disclaimer]
      properties:
        suggested_type: { $ref: '#/components/schemas/ThemeType' }
        disclaimer: { type: string }
    SimulationType:
      type: string
      enum: [protanopia, deuteranopia, tritanopia]
      description: Familia de deficiencia cromatica usada por las matrices de Machado.
    SimulationSeverity:
      type: number
      format: double
      minimum: 0
      maximum: 1
      default: 1
      description: 0 conserva el color original y 1 aplica la simulacion completa.
    SimulateRequest:
      type: object
      additionalProperties: false
      required: [hex, type]
      properties:
        hex: { $ref: '#/components/schemas/HexColor' }
        type: { $ref: '#/components/schemas/SimulationType' }
        severity: { $ref: '#/components/schemas/SimulationSeverity' }
    SimulateResponse:
      type: object
      additionalProperties: false
      required: [original, simulated, type, severity, model]
      properties:
        original: { $ref: '#/components/schemas/HexColor' }
        simulated: { $ref: '#/components/schemas/HexColor' }
        type: { $ref: '#/components/schemas/SimulationType' }
        severity: { $ref: '#/components/schemas/SimulationSeverity' }
        model:
          type: string
          enum: [machado-2009]
    SimulatedColor:
      type: object
      additionalProperties: false
      required: [original, simulated]
      properties:
        original: { $ref: '#/components/schemas/HexColor' }
        simulated: { $ref: '#/components/schemas/HexColor' }
    SimulateBatchRequest:
      type: object
      additionalProperties: false
      required: [colors, type]
      properties:
        colors:
          type: array
          minItems: 1
          maxItems: 256
          items: { $ref: '#/components/schemas/HexColor' }
        type: { $ref: '#/components/schemas/SimulationType' }
        severity: { $ref: '#/components/schemas/SimulationSeverity' }
    SimulateBatchResponse:
      type: object
      additionalProperties: false
      required: [type, severity, model, results]
      properties:
        type: { $ref: '#/components/schemas/SimulationType' }
        severity: { $ref: '#/components/schemas/SimulationSeverity' }
        model:
          type: string
          enum: [machado-2009]
        results:
          type: array
          items: { $ref: '#/components/schemas/SimulatedColor' }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          enum: [invalid_type, invalid_parameter, invalid_request, invalid_palette, invalid_color, not_found, method_not_allowed, rate_limited, timeout, internal_server_error, https_required]
        message: { type: string }
  responses:
    BadRequest:
      description: Solicitud invalida.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    MethodNotAllowed:
      description: Metodo no permitido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    HttpsRequired:
      description: HTTPS es obligatorio cuando el servicio esta en produccion.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Limite publico de solicitudes excedido.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Timeout:
      description: Tiempo de ejecucion excedido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Error interno sin stack trace.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }